refactor: pickerComponents 분리
